+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## Unreleased
+## 1.4.0
 - Added ignore plugin loading list option.
+- Added logging of SteamNetworkingSockets errors due to unable to send message.
+- Added NoNagle to some RPC delivery flags.
 
 ## 1.3.0
 - Added log and exception handler of null AudioSource in AudioReverbTrigger.